feature :add task manager

+++ task_manager/task_handler.go
@@ -2,34 +2,53 @@ package task_manager
 
 import (
 	"context"
+	"encoding/json"
+	"fmt"
+	"github.com/pkg/errors"
 	clientv3 "go.etcd.io/etcd/client/v3"
 	"golib/concurrency/gopool"
 	"golib/etcd_helper"
 	"golib/logger"
 	"reflect"
+	"runtime"
 	"sync"
 	"time"
 )
 
-type TaskFuncType func(ctx context.Context, params interface{}) (err error)
-
 type TaskHandler struct {
-	Name       TaskHandlerName
-	Config     *TaskConfig
-	TaskFunc   TaskFuncType
-	ParamsType reflect.Type
+	Name     TaskHandlerName
+	Config   *TaskConfig
+	TaskFunc interface{}
 
 	runningTasks sync.Map
 
 	taskHandlerGoPool gopool.Pool // 控制handler的并发度
 }
 
-func NewTaskHandler(name TaskHandlerName, config *TaskConfig, taskFunc TaskFuncType, paramsType interface{}) *TaskHandler {
+func NewTaskHandler(name TaskHandlerName, config *TaskConfig, taskFunc interface{}) *TaskHandler {
+	t := reflect.TypeOf(taskFunc)
+	if t.Kind() != reflect.Func {
+		panic(fmt.Sprintf("%s 非法方法", runtime.FuncForPC(reflect.ValueOf(taskFunc).Pointer()).Name()))
+	}
+	inNum := t.NumIn()
+	if inNum != 2 {
+		panic(fmt.Sprintf("%s 非法方法入参", runtime.FuncForPC(reflect.ValueOf(taskFunc).Pointer()).Name()))
+	}
+	if t.In(0).Name() != "Context" {
+		panic(fmt.Sprintf("%s 非法方法入参", runtime.FuncForPC(reflect.ValueOf(taskFunc).Pointer()).Name()))
+	}
+	outNum := t.NumOut()
+	if outNum != 1 {
+		panic(fmt.Sprintf("%s 非法方法出参", runtime.FuncForPC(reflect.ValueOf(taskFunc).Pointer()).Name()))
+	}
+	if t.Out(0).Name() != "error" {
+		panic(fmt.Sprintf("%s 方法出参应为error", runtime.FuncForPC(reflect.ValueOf(taskFunc).Pointer()).Name()))
+	}
+
 	return &TaskHandler{
 		Name:              name,
 		Config:            config,
 		TaskFunc:          taskFunc,
-		ParamsType:        reflect.TypeOf(paramsType),
 		taskHandlerGoPool: gopool.NewPool("taskHandlerGoPool", gopool.NewConfig(gopool.WithConfigCap(100))),
 	}
 }
@@ -48,7 +67,33 @@ func (handler *TaskHandler) exec(ctx context.Context, task *Task) (err error) {
 			return
 		default:
 		}
-		errChan <- handler.TaskFunc(ctx, task.Params)
+		reValues := make([]reflect.Value, 0)
+		reValueCtx := reflect.ValueOf(ctx)
+		reValues = append(reValues, reValueCtx)
+
+		fT := reflect.TypeOf(handler.TaskFunc)
+		v := reflect.New(fT.In(1))
+		paramBytes, err := json.Marshal(task.Params)
+		if err != nil {
+			errChan <- err
+		}
+		err = json.Unmarshal(paramBytes, v.Interface())
+		if err != nil {
+			errChan <- err
+		}
+		reValues = append(reValues, v.Elem())
+
+		fv := reflect.ValueOf(handler.TaskFunc)
+		resVal := fv.Call(reValues)
+		//NewTaskHandler做了前置判断
+		if len(resVal) == 1 {
+			if !resVal[0].IsNil() {
+				errChan <- errors.New(fmt.Sprintf("%v", resVal[0]))
+			} else {
+				errChan <- nil
+			}
+		}
+
 	})
 	select {
 	case <-ctx.Done():
@@ -59,7 +104,7 @@ func (handler *TaskHandler) exec(ctx context.Context, task *Task) (err error) {
 		if task.ExecCount >= handler.Config.Retry {
 			task.markFail()
 		} else {
-			task.markPending(handler.Config.DelayTime)
+			task.markPending()
 		}
 		return err
 	}
@@ -69,11 +114,32 @@ func (handler *TaskHandler) exec(ctx context.Context, task *Task) (err error) {
 
 func (handler *TaskHandler) handleTask(ctx context.Context, task *Task) {
 	ctx = logger.CtxWithTraceId(ctx, task.TraceId)
+	task.ctx, task.cancelCtx = context.WithCancel(ctx)
 	switch task.Status {
 	case pending:
 		handler.taskHandlerGoPool.Go(func() {
 			handler.handlePendingTask(ctx, task)
 		})
+	case del:
+		value, ok := handler.runningTasks.Load(task.getTaskKey().String())
+		if !ok {
+			return
+		}
+		runningTask, ok := value.(*Task)
+		if !ok {
+			return
+		}
+		// 删除正在内存中跑的任务，再去删除ETCD中的任务
+		runningTask.cancelCtx()
+		handler.runningTasks.Delete(task.getTaskKey().String())
+		_, err := etcd_helper.Context(ctx).
+			Txn(ctx).
+			If(clientv3.Compare(clientv3.Version(task.getTaskKey().String()), "=", task.TaskVersion)).
+			Then(clientv3.OpDelete(task.getTaskKey().String())).
+			Commit()
+		if err != nil {
+			logger.CtxErrorf(ctx, "delete task fail: %+v", err)
+		}
 	}
 }
 
@@ -94,10 +160,12 @@ func (handler *TaskHandler) handlePendingTask(ctx context.Context, task *Task) {
 	after := task.NextExecuteTime.Sub(time.Now())
 	select {
 	case <-time.After(after):
+	case <-task.ctx.Done(): // 任务的context被取消则直接结束
+		return
 	}
-	err = handler.exec(ctx, task)
+	err = handler.exec(task.ctx, task)
 	if err != nil {
-		logger.CtxErrorf(ctx, "handler.exec err: %s", err.Error())
+		logger.CtxErrorf(ctx, "handler.exec err: %s, task:%+v", err.Error(), task)
 		return
 	}
 }
@@ -121,6 +189,12 @@ func (handler *TaskHandler) preemptTask(ctx context.Context, task *Task) (err er
 }
 
 func (handler *TaskHandler) completeTask(ctx context.Context, task *Task) error {
+	_, ok := handler.runningTasks.Load(task.getTaskKey().String())
+	if !ok {
+		// 任务已经被删除，被强制中断，不需要做任何操作
+		return nil
+	}
+
 	if task.Status == success {
 		deleteResponse, err := etcd_helper.Delete(ctx, task.getTaskKey().String())
 		if err != nil {
@@ -155,7 +229,9 @@ func (handler *TaskHandler) interruptRunningTask() error {
 			return true
 		}
 		ctx := etcd_helper.BindContext(logger.CtxWithTraceId(context.Background(), task.TraceId))
-		task.markPending(0)
+		task.markPending()
+		// 被打断，不算执行次数
+		task.ExecCount--
 		_, err := etcd_helper.Put(ctx, task.getTaskKey().String(), task.encode())
 		if err != nil {
 			logger.CtxErrorf(ctx, "close interruptRunningTask fail: %+v", err)
